# Avoid calling go_path rule in source mode