--- a/tests/core/gomock/BUILD.bazel
+++ b/tests/core/gomock/BUILD.bazel
@@ -1,0 +1,33 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test", "gomock")
+
+go_library(
+    name = "client",
+    srcs = [
+        "client.go",
+    ],
+    importpath = "uber.com/client",
+    visibility = ["//visibility:public"],
+    deps = [
+        "@go_googleapis//google/bytestream:bytestream_go_proto",
+        "@org_golang_google_grpc//:grpc",
+    ],
+)
+
+gomock(
+    name = "mocks",
+    out = "client_mock.go",
+    interfaces = ["Client"],
+    library = ":client",
+    package = "client",
+    source = "client.go",
+)
+
+go_test(
+    name = "client_test",
+    srcs = [
+        "client_mock.go",
+        "client_test.go",
+    ],
+    embed = [":client"],
+    deps = ["@com_github_golang_mock//gomock"],
+)