--- a/tests/core/gomock/client.go
+++ b/tests/core/gomock/client.go
@@ -1,0 +1,10 @@
+package client
+
+import (
+	"google.golang.org/genproto/googleapis/bytestream"
+	"google.golang.org/grpc"
+)
+
+type Client interface {
+	Connect(grpc.ClientConnInterface) *bytestream.ByteStreamClient
+}