--- a/tests/core/gomock/client_test.go
+++ b/tests/core/gomock/client_test.go
@@ -1,0 +1,3 @@
+package client
+
+var _ Client = (*MockClient)(nil)