… third_party/com_github_golang_mock-gazelle.patch