**What type of PR is this?**
Performance optimization. This is also a breaking change from original `gomock` rule implementation, because it changes the types of `aux_files` attribute.

**What does this PR do? Why is it needed?**
Running mockgen in source mode does not require all its dependencies in GOPATH. It only need the source file and some aux files. We can copy them to form a GOPATH structure without calling `go_path` rule. 

Before this patch:
```
Critical path (1.199 s):
       Time Percentage   Description
    0.46 ms    0.04%   action 'Writing file mocks_gomock_gopath~manifest'
     250 ms   20.86%   action 'GoPath mocks_gomock_gopath'
     949 ms   79.10%   action 'Action mocks.go'
```
After this patch:

```
Critical path (891 ms):
       Time Percentage   Description
    24.1 ms    2.70%   action 'Action gopath/src/uber.com/gomock/client.go'
     867 ms   97.30%   action 'Action mocks.go'
```

The experiment can be reproduced in the following workspace:

1. create a workspace with the following files:
```
-- BUILD.bazel --
load("@io_bazel_rules_go//go:def.bzl", "go_library")
load("@io_bazel_rules_go//extras:gomock.bzl", "gomock")
load("@bazel_gazelle//:def.bzl", "gazelle")

# gazelle:prefix uber.com/gomock
gazelle(name = "gazelle")

gazelle(
    name = "gazelle-update-repos",
    args = [
        "-from_file=go.mod",
        "-prune",
    ],
    command = "update-repos",
)

go_library(
    name = "client",
    srcs = ["client.go"],
    importpath = "uber.com/gomock",
    visibility = ["//visibility:public"],
    deps = [
        "@go_googleapis//google/bytestream:bytestream_go_proto",
        "@org_golang_google_grpc//:go_default_library",
    ],
)

# gazelle:exclude mocks.go
gomock(
    name = "mocks",
    out = "mocks.go",
    library = ":client",
    package = "gomock",
    source = "client.go",
    interfaces = ["Client"],
)
-- client.go --
package gomock

import (
	_ "github.com/mattn/go-sqlite3"
	_ "google.golang.org/genproto/googleapis/bytestream"
	_ "google.golang.org/grpc"
)

type Client interface {
}
-- go.mod --
module uber.com/gomock

go 1.18

require (
	github.com/mattn/go-sqlite3 v1.14.12
	google.golang.org/genproto v0.0.0-20220414192740-2d67ff6cf2b4
	google.golang.org/grpc v1.45.0
)

require (
	github.com/golang/protobuf v1.5.2 // indirect
	github.com/google/go-cmp v0.5.7 // indirect
	golang.org/x/net v0.0.0-20211015210444-4f30a5c0130f // indirect
	golang.org/x/sys v0.0.0-20220319134239-a9b59b0215f8 // indirect
	golang.org/x/text v0.3.7 // indirect
	google.golang.org/protobuf v1.28.0 // indirect
)
-- WORKSPACE --
load("@bazel_tools//tools/build_defs/repo:http.bzl", "http_archive")

http_archive(
    name = "io_bazel_rules_go",
    sha256 = "6c3bc93b2f14db41c2ef53388700fb95cbf7005b7c94e98f5d8bca0748110062",
    strip_prefix = "rules_go-926c47abbb2dd059a4fcd3f810005167cc62ed1c",
    urls = [
        "https://github.com/bazelbuild/rules_go/archive/926c47abbb2dd059a4fcd3f810005167cc62ed1c.zip",
    ],
)

http_archive(
    name = "bazel_gazelle",
    sha256 = "de69a09dc70417580aabf20a28619bb3ef60d038470c7cf8442fafcf627c21cb",
    urls = [
        "https://mirror.bazel.build/github.com/bazelbuild/bazel-gazelle/releases/download/v0.24.0/bazel-gazelle-v0.24.0.tar.gz",
        "https://github.com/bazelbuild/bazel-gazelle/releases/download/v0.24.0/bazel-gazelle-v0.24.0.tar.gz",
    ],
)

load("@io_bazel_rules_go//go:deps.bzl", "go_register_toolchains", "go_rules_dependencies")
load("@bazel_gazelle//:deps.bzl", "gazelle_dependencies", "go_repository")

go_rules_dependencies()

go_register_toolchains(version = "1.18")

gazelle_dependencies()

http_archive(
    name = "com_google_protobuf",
    sha256 = "d0f5f605d0d656007ce6c8b5a82df3037e1d8fe8b121ed42e536f569dec16113",
    strip_prefix = "protobuf-3.14.0",
    urls = [
        "https://mirror.bazel.build/github.com/protocolbuffers/protobuf/archive/v3.14.0.tar.gz",
        "https://github.com/protocolbuffers/protobuf/archive/v3.14.0.tar.gz",
    ],
)

load("@com_google_protobuf//:protobuf_deps.bzl", "protobuf_deps")

protobuf_deps()

go_repository(
    name = "com_github_golang_mock",
    build_file_generation = "on",
    importpath = "github.com/golang/mock",
    sum = "h1:G5FRp8JnTd7RQH5kemVNlMeyXQAztQ3mOWV95KxsXH8=",
    version = "v1.1.1",
)
```
2. run `go mod tidy`
3. run `bazel run //:gazelle-update-repos`
4. run `bazel build --profile=/tmp/mocks.gz //:mocks`
5. run `bazel analyze-profile /tmp/mocks.gz`