@achew22 2022-04-16
This looks good to me. Is there any chance we could use add a test? Even just something that used it would provide a lot of confidence for me. We could also add a test in https://github.com/bazelbuild/rules_go/tree/master/tests/integration to confirm it if you're feeling it.

Also, what would you think about explicitly adding some docs onto this that say the API on the gomock rules isn't stable yet? Then we can make some changes, wait for it to solidify and then promote it?