@linzhp 2022-04-17
Good ideas. I added a unit test, instead of an integration test, to generate mocks using the `gomock` rule, and verify the output. Unit tests should be powerful enough in this case, and they are much faster.